Refactor for better usability
Several imporvements have been implemented by code refactoring.
The library is still under development and it is benefitting
from the implementation of a hand tracking algorithm.
In particular, it has been noticed that "FilteringFunction" classes
need to be divided into single per-operation/task classes
to significantly improve code re-usability.
From this commit onward, changes will be made to ease
the development of the hand-tracking algorithm.

Another branch will be started to address major improvements
to the library.