Add new implementation of FilteringAlgorithm
The filters now expose a new interface.
The runFilter() method is changed to filteringStep() and implements a single filtering step.
Filter initialization must be implemented by overriding initialization().
To stop a filter (mostly for simulation purposes), runCondition() must be override.
The following new method can be used and are defined as final to block any possible override:
 - prepare(): initialize the filter and wait for run() call;
 - run(): run the filter;
 - wait(): wait the filter until runCondition() returns false;
 - reset(): reset the filter, i.e. initialize + run;
 - reboot(): reset the filter but wait for run() function call again;
 - teardown(): stop and close the filter.