New PF prediction implementation
The new implementation uses both the Strategy and Decroator pattern.
In particular, the pointer for the StateModel is in the concrete class
of the ParticleFilterPrediction interface, and the decorators can access
it via a getter function. In this way, the strategy changes the
behaviour of the concrete implementation of the prediction, while the
decorator changes the behaviour of the concrete class while being
agnostic to the strategy.